apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3.0
supabase>=2.0.0,<3.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
trafilatura>=1.6.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
